B_phy: 6차 과제

+++ likelion_11th/main/views.py
@@ -18,21 +18,27 @@ def new(request): #앞의 new.html을 띄우는 함수
     return render(request, 'main/new.html')
 
 def create(request): #포스트 생성(CRUD 중 C)
-    new_post = Post()
-    new_post.title = request.POST['title']
-    new_post.writer = request.POST['writer']
-    new_post.address = request.POST['address']
-    new_post.weather = request.POST['weather']
-    new_post.pub_date = timezone.now()
-    new_post.body = request.POST['body']
-    new_post.image = request.FILES.get('image')
+    if request.user.is_authenticated:
+        new_post = Post()
+        new_post.title = request.POST['title']
+        new_post.writer = request.POST['writer']
+        new_post.address = request.POST['address']
+        new_post.weather = request.POST['weather']
+        new_post.pub_date = timezone.now()
+        new_post.body = request.POST['body']
+        new_post.image = request.FILES.get('image')
+        new_post.user = request.user
+        
+        new_post.save()
+        
+        return redirect('main:detail', new_post.id) #새로 생성한 post id와함께 detail 페이지로 이동
     
-    new_post.save()
-    
-    return redirect('main:detail', new_post.id) #새로 생성한 post id와함께 detail 페이지로 이동
+    else:
+        return redirect('main:mainpage')
 
 def detail(request, id): #id에 원하는 게시글의 id 값을 넣어 detail 함수를 실행
     post = get_object_or_404(Post, pk = id) #Post와 id를 받아서 전송 or 오류표시
+    
     return render(request, 'main/detail.html', {'post':post}) # id에 부합하는 게시물 1개씩 관리(detail 페이지)
     # pk(Primary Key): 각 객체를 구분해주는 키 값
     
@@ -41,19 +47,25 @@ def edit(request, id):
     return render(request, 'main/edit.html', {'post': edit_post})
 
 def update(request, id):
-    update_post = Post.objects.get(id=id)
-    update_post.title = request.POST['title']
-    update_post.writer = request.POST['writer']
-    update_post.address = request.POST['address']
-    update_post.weather = request.POST['weather']
-    update_post.pub_date = timezone.now()
-    update_post.body = request.POST['body']
-    update_post.image = request.FILES.get('image', update_post.image) #파일 수정시에도 기존 기록 남게
-    
-    update_post.save()
-    return redirect('main:detail', update_post.id)
+    if request.user.is_authenticated:
+        update_post = Post.objects.get(id=id)
+        update_post.title = request.POST['title']
+        update_post.writer = request.POST['writer']
+        update_post.address = request.POST['address']
+        update_post.weather = request.POST['weather']
+        update_post.pub_date = timezone.now()
+        update_post.body = request.POST['body']
+        update_post.image = request.FILES.get('image', update_post.image) #파일 수정시에도 기존 기록 남게
+        
+        update_post.save()
+        return redirect('main:detail', update_post.id)
+    else:
+        return redirect('main:mainpage')
 
 def delete(request, id):
-    delete_post = Post.objects.get(id=id)
-    delete_post.delete()
-    return redirect('main:mainpage')
+    if request.user.is_authenticated:
+        delete_post = Post.objects.get(id=id)
+        delete_post.delete()
+        return redirect('main:mainpage')
+    else:
+        return redirect('main:mainpage')
